src/main/webapp/app/entities/transaction-types/route/transaction-types-routing-resolve.service.spec.ts
import { TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRouteSnapshot, ActivatedRoute, Router, convertToParamMap } from '@angular/router';
import { RouterTestingModule } from '@angular/router/testing';
import { of } from 'rxjs';

import { ITransactionTypes, TransactionTypes } from '../transaction-types.model';
import { TransactionTypesService } from '../service/transaction-types.service';

import { TransactionTypesRoutingResolveService } from './transaction-types-routing-resolve.service';

describe('TransactionTypes routing resolve service', () => {
  let mockRouter: Router;
  let mockActivatedRouteSnapshot: ActivatedRouteSnapshot;
  let routingResolveService: TransactionTypesRoutingResolveService;
  let service: TransactionTypesService;
  let resultTransactionTypes: ITransactionTypes | undefined;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, RouterTestingModule.withRoutes([])],
      providers: [
        {
          provide: ActivatedRoute,
          useValue: {
            snapshot: {
              paramMap: convertToParamMap({}),
            },
          },
        },
      ],
    });
    mockRouter = TestBed.inject(Router);
    jest.spyOn(mockRouter, 'navigate').mockImplementation(() => Promise.resolve(true));
    mockActivatedRouteSnapshot = TestBed.inject(ActivatedRoute).snapshot;
    routingResolveService = TestBed.inject(TransactionTypesRoutingResolveService);
    service = TestBed.inject(TransactionTypesService);
    resultTransactionTypes = undefined;
  });

  describe('resolve', () => {
    it('should return ITransactionTypes returned by find', () => {
      // GIVEN
      service.find = jest.fn(id => of(new HttpResponse({ body: { id } })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultTransactionTypes = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultTransactionTypes).toEqual({ id: 123 });
    });

    it('should return new ITransactionTypes if id is not provided', () => {
      // GIVEN
      service.find = jest.fn();
      mockActivatedRouteSnapshot.params = {};

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultTransactionTypes = result;
      });

      // THEN
      expect(service.find).not.toBeCalled();
      expect(resultTransactionTypes).toEqual(new TransactionTypes());
    });

    it('should route to 404 page if data not found in server', () => {
      // GIVEN
      jest.spyOn(service, 'find').mockReturnValue(of(new HttpResponse({ body: null as unknown as TransactionTypes })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultTransactionTypes = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultTransactionTypes).toEqual(undefined);
      expect(mockRouter.navigate).toHaveBeenCalledWith(['404']);
    });
  });
});
